redesign pool management
it compiles but has no guarantees of not blowing up in your face

- do not create all pool workers when a pool is created
- use leases to mark liveness of workers. discard workers after their liveness is up, allowing new ones to be created
- use metadata to tag ownership of workers.
- allow workers to persist pool machine restarts, and recapture workers on startup.
- implement cleaner that eliminates any workers that no longer have leases

+++ machines/lease.go
@@ -19,7 +19,7 @@ type LeaseResp struct {
 
 type LeaseData struct {
 	Nonce     string `json:"nonce"`
-	ExpiresAt int    `json:"expires_at"`
+	ExpiresAt int64  `json:"expires_at"`
 	Owner     string `json:"owner"`
 	Descr     string `json:"description"`
 	Version   string `json:"version"`
